Add distinguish between R and other devices
